# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/products/all": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "All Products - Publicly Accessible",
                "description": "All Products - Publicly Accessible",
                "operationId": "indexAll",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "perPage, eg; 20",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All Products - Publicly Accessible"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Show Product Details",
                "description": "Show Product Details",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Product Details"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update Product",
                "description": "Update Product",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Product 1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 10120
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update Product"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete Product",
                "description": "Delete Product",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Product"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/restaurants": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "List of Restaurants",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "0ef65917394f8eb77311d1d085ff31e4",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/restaurants/all": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "All Restaurants - Publicly Accessible",
                "description": "All Restaurants - Publicly Accessible",
                "operationId": "0e390501c6147527b627735cf517bcfc",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "perPage, eg; 20",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All Restaurants - Publicly Accessible"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/restaurants/search": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "All Restaurants - Publicly Accessible",
                "description": "All Restaurants - Publicly Accessible",
                "operationId": "b54700b170fad1c9280cb8a2cc00eef3",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "perPage, eg; 20",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search, eg; Test",
                        "schema": {
                            "type": "string"
                        },
                        "example": "Test"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All Restaurants - Publicly Accessible"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/restaurants/store": {
            "post": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Create a Restaurant",
                "operationId": "9d2cb3c8a9c036f238ac4c9b5d65a8b0",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "description": "Name of the restaurant",
                                                "type": "string"
                                            },
                                            "Location": {
                                                "description": "Location of the restaurant",
                                                "type": "string"
                                            },
                                            "Address": {
                                                "description": "Adress of the restaurant's address",
                                                "type": "string"
                                            },
                                            "Phone Number": {
                                                "description": "Phone Number of the restaurant",
                                                "type": "string"
                                            },
                                            "Description": {
                                                "description": "City in which the restaurant is located",
                                                "type": "string"
                                            },
                                            "user_id": {
                                                "description": "ID of the user who owns the restaurant",
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "KFC",
                                    "location": "LA",
                                    "address": "123 ST PARIS",
                                    "telephone": "123-456-7890",
                                    "phone_number": "555-555-5555",
                                    "description": "Fried chicken specialist",
                                    "user_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "KFC"
                                        },
                                        "Location": {
                                            "type": "string",
                                            "example": "LA"
                                        },
                                        "Address": {
                                            "type": "string",
                                            "example": "123 ST PARIS"
                                        },
                                        "Phone Number": {
                                            "type": "string",
                                            "example": "555-555-5555"
                                        },
                                        "Description": {
                                            "type": "string",
                                            "example": "Fried chicken specialist"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-01-15T15:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-01-15T15:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/restaurants/show/{id}": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Get a Restaurant",
                "operationId": "9c1f8a2e091fedf8b9dfa805da8520f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "KFC"
                                        },
                                        "Location": {
                                            "type": "string",
                                            "example": "LA"
                                        },
                                        "Address": {
                                            "type": "string",
                                            "example": "123 ST PARIS"
                                        },
                                        "Phone Number": {
                                            "type": "string",
                                            "example": "555-555-5555"
                                        },
                                        "Description": {
                                            "type": "string",
                                            "example": "Fried chicken specialist"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-01-15T15:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-01-15T15:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/restaurants/delete/{id}": {
            "delete": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Delete a Restaurant",
                "operationId": "f6781cddca299991a0feb6d9cdca837d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "success"
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}